Refactor to remove `validator_node`
The `validator_node` create was only being used to hold the
NodeImplementation binding Espresso to HotShot. It's confusing name,
and the waste of having an entire crate just for one, short file
that is only referenced once (from the `validator` crate) meant it
was better just to move the `NodeImplementation` to the `validator`
crate and remove the `validator_node` crate.

I also removed an ancient wallet fuzz test that referenced
`validator_node`. This test wasn't being used, probably wouldn't
work anyway, and we have plenty of wallet tests elsewhere.
