+++ client/Cargo.toml
@@ -69,7 +69,6 @@ tempdir = "0.3.7"
 toml = "0.5"
 tracing = "0.1.35"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-validator_node = { path = "../validator_node/" }
 
 [dev-dependencies]
 seahorse = { git = "https://github.com/EspressoSystems/seahorse.git", features = ["testing"] }
