--- core/fuzz/Cargo.toml
@@ -1,44 +1,0 @@
-# Copyright (c) 2022 Espresso Systems (espressosys.com)
-# This file is part of the Espresso library.
-#
-# This program is free software: you can redistribute it and/or modify it under the terms of the GNU
-# General Public License as published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-# This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
-# even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
-# General Public License for more details.
-# You should have received a copy of the GNU General Public License along with this program. If not,
-# see <https://www.gnu.org/licenses/>.
-
-[package]
-name = "espresso-core-fuzz"
-version = "0.0.0"
-authors = ["Jeb Bearer <jeb.bearer@translucence.net>"]
-publish = false
-edition = "2018"
-
-[package.metadata]
-cargo-fuzz = true
-
-[dependencies]
-libfuzzer-sys = "0.4"
-async-std = { version = "1.10.0", features = ["unstable", "attributes"] }
-
-[dependencies.validator_node]
-path = "../../validator_node"
-
-[dependencies.espresso-core]
-path = ".."
-
-# Prevent this from interfering with workspaces
-[workspace]
-members = ["."]
-
-[profile.release.package."*"]
-opt-level = 3
-
-[[bin]]
-name = "multixfr_keystore"
-path = "fuzz_targets/multixfr_keystore.rs"
-test = false
-doc = false
