+++ validator/Cargo.toml
@@ -74,7 +74,6 @@ tracing = "0.1.35"
 tracing-distributed = "0.3.1"
 tracing-futures = "0.2"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-validator_node = { path = "../validator_node/" }
 
 [features]
 slow-tests = []
